requirements.txt
# LLM + embeddings
openai>=1.30.0
anthropic>=0.25.0
tiktoken>=0.7.0

# RAG framework (baseline)
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# Vector store
chromadb>=0.5.0

# Hybrid search
rank-bm25>=0.2.2

# Reranking
sentence-transformers>=3.0.0

# Document loading
pymupdf>=1.24.0         # PDF parsing
python-pptx>=0.6.23    # PowerPoint slides
python-docx>=1.1.0     # Word docs
unstructured>=0.14.0   # General doc loader fallback

# Evaluation
ragas>=0.1.0
rouge-score>=0.1.2
nltk>=3.8.0

# App
streamlit>=1.35.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
tqdm>=4.66.0
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.26.0

# Vision / image rendering (for PPTX slide rendering)
pdf2image>=1.17.0          # fallback slide renderer (needs poppler)
